Bug correction
I thought I corrected the bug the last time but I didn't.
The problem was that at the end of a bloc I assumed a block took 70us (wtf!?) to finish but it was not the case. So when waiting at the end of the loop I waited negative time casted to unsigned value and it was a huge amount.

I corrected this but now I cannot execute all the clocks without burning my laptop (I was abble using ~16% of my cpu before correction)