Improving emulation stop
Debugger cleanup sequence check if the debugger is in an initialized state or not, and only clean if the debugger is in an initialized state.
This is to make sure that the fifo are closed when the emulation is stopped by a signal. Signal handler only set the continue value to false,
but if the debugger is in a read or write state and is blocked in it, the cleanup might never be called, the fifo never closed and the debugger
never ended.

Cleanup is now made when signal handler is called or at the end of the emulation loop